Fix [RB-5]
Created `PowerGPA::APIClient::IncorrectCredentialsError` error class.
Within the `error` block, we now check for this specific error message
and will return a special error message if invalid credentials are
entered. We will also, from this point forward, exclude this error from
being reported in Rollbar, due to the fact that we can do nothing about
it.